+++ src/index.ts
@@ -1,13 +1,13 @@
 import { registerPlugin } from '@capacitor/core';
 
-import type { FoldablePlugin } from './definitions';
+import type { FoldablePlugin, FoldablePolyfillOptions } from './definitions';
 import { install } from './polyfill';
 
 const Foldable = registerPlugin<FoldablePlugin>('Foldable', {
   web: () => import('./web').then((m) => new m.FoldableWeb()),
 });
 
-const installFoldablePolyfill = (): Promise<void> => install(Foldable);
+const installFoldablePolyfill = (options?: FoldablePolyfillOptions): Promise<void> => install(Foldable, options);
 
 export * from './definitions';
 export { Foldable, installFoldablePolyfill };